feat:작업하기페이지 사이드바 기능 구현 상단 버튼 CSS수정
