+++ src/app/(provider)/(root)/challenges/[challengeId]/challengeTask/challengeTask.module.css
@@ -20,14 +20,35 @@
 .exitButton {
   width: 81px;
   height: 40px;
+  background-color: #ffe7e7;
+  font-size: 16px;
+  font-weight: 600;
+  color: #f24744;
+  border-radius: 12px;
+  border: none;
+  cursor: pointer;
 }
 .saveDraftButton {
   width: 90px;
   height: 40px;
+  background-color: white;
+  color: #262626;
+  font-size: 16px;
+  font-weight: 600;
+  border-radius: 12px;
+  border: 1.5px solid #262626;
+  cursor: pointer;
 }
 .submitButton {
   width: 90px;
   height: 40px;
+  background-color: #262626;
+  color: white;
+  font-size: 16px;
+  font-weight: 600;
+  border-radius: 12px;
+  border: none;
+  cursor: pointer;
 }
 
 .horizontalLine {
