+++ src/app/(provider)/(root)/challenges/[challengeId]/challengeTask/sideBar.module.css
@@ -1,0 +1,40 @@
+.container {
+  background-color: #e3ecf1;
+}
+.sidebar {
+  background-color: #e3ecf1;
+  border-left: 4px solid #202020;
+  position: fixed;
+  top: 0;
+  bottom: 0;
+  right: 0;
+  transition: 0.4s ease;
+  color: #202020;
+  height: 100%;
+  z-index: 99;
+}
+.button {
+  position: relative;
+  left: -50px;
+  top: 136px;
+  width: 52px;
+  height: 99px;
+  z-index: 99;
+  transition: 0.8s ease;
+  border: 2px solid #f5f5f5;
+  border-left: 24;
+  overflow: hidden;
+}
+.openBtn {
+  width: 100%;
+  height: 100%;
+}
+.content {
+  padding: 40px 40px 0 20px;
+  position: relative;
+  width: 100%;
+}
+.icon {
+  margin: 0;
+  color: #202020;
+}
