+++ src/app/(provider)/(root)/my-challenges/myChallenge.module.css
@@ -48,7 +48,7 @@
 .header_btns button {
   width: 162px;
   height: 53px;
-  background-color: white;
+  background-color: #fafafa;
 
   border: none;
