+++ src/components/Button/FilterButton.module.css
@@ -11,6 +11,7 @@
   justify-content: space-between;
   align-items: center;
   border: 1px solid #d4d4d4;
+  background-color: white;
   border-radius: 32px;
   padding: 8px 12px;
   box-sizing: border-box;
