+++ src/components/Headers/Header.module.css
@@ -15,6 +15,11 @@
   margin-left: 10rem;
 }
 
+.logo a {
+  display: flex;
+  align-items: center;
+}
+
 .nav {
   display: flex;
 }
@@ -36,6 +41,7 @@
 .icons img {
   margin-left: 0.5rem;
   cursor: pointer;
+  vertical-align: bottom;
 }
 
 .menuContainer {
